Update index.md
An update to the note under "Testing — create some items" is part of my solution to #27368 in this (mdn/content) repo. The other part is in another repo, /mdn/express-locallibrary-tutorial. I don't know if it's possible to link the two, but merging them at different times won't cause any problems.